server/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
Authlib==1.3.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
distro==1.9.0
dnspython==2.5.0
exceptiongroup==1.2.0
Flask==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
openai==1.10.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
